moved get progress into current challenge so it gets the current challenge data in the progress bar
